tests/submission/Infraware/Forms/contents/Forms/textarea_labels.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_labels</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
		<p><label>Full name:<label>(name)<textarea id='textarea_id1'></textarea></label></label></p>
		<p><label>Age:<textarea id='textarea_id2'></textarea></label></p>
	</form>
	<script>

		var textarea1 = document.getElementById("textarea_id1");
		var textarea2 = document.getElementById("textarea_id2");

		if (typeof(textarea1.labels) == "object") {
			if (textarea1.labels.length == 2 && textarea2.labels.length == 1) {
				test( function() {
					assert_true(true, "labels attribute is correct.");
				});
			} else {
				test( function() {
					assert_unreached("labels attribute is not correct.");
				});
			}
		} else {
			test( function() {
				assert_unreached("labels attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
